requirements.txt
MetaTrader5>=5.0.45
pandas>=2.0.0
pandas-ta>=0.3.14b
numpy>=1.24.0
loguru>=0.7.0
APScheduler>=3.10.0
python-dotenv>=1.0.0
pytest>=7.4.0
streamlit>=1.30.0
plotly>=5.18.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
yfinance>=0.2.0
torch>=2.0.0
anthropic>=0.40.0
scikit-optimize>=0.9.0
